# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/v1/auth/login": {
      "post": {
        "operationId": "Auth_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "example": {
                "email": "admin@orbit.sa",
                "password": "Orbit@2026"
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "email": "admin@orbit.sa",
                    "password": "Orbit@2026"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "tags": [
          "Authentication"
        ],
        "summary": "Create / execute authentication",
        "description": "Create / execute operation in the Authentication module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/auth/refresh": {
      "post": {
        "operationId": "Auth_refresh",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              },
              "example": {
                "refreshToken": "eyJ..."
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "refreshToken": "eyJ..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "tags": [
          "Authentication"
        ],
        "summary": "Create / execute authentication",
        "description": "Create / execute operation in the Authentication module. Organization scope is enforced from the authenticated session."
      }
    },
    "/v1/auth/logout": {
      "post": {
        "operationId": "Auth_logout",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshDto"
              },
              "example": {
                "refreshToken": "eyJ..."
              },
              "examples": {
                "complete": {
                  "summary": "Complete request example",
                  "value": {
                    "refreshToken": "eyJ..."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "tags": [
          "Authentication"
        ],
        "summary": "Create / execute authentication",
        "description": "Create / execute operation in the Authentication module. Organization scope is enforced from the authenticated session.",
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ]
      }
    },
    "/v1/auth/me": {
      "get": {
        "operationId": "Auth_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Validation failed"
          },
          "401": {
            "description": "Missing, invalid or expired authentication"
          },
          "403": {
            "description": "Authenticated role is not permitted"
          },
          "404": {
            "description": "Resource not found in the current organization"
          }
        },
        "security": [
          {
            "bearerAuth": []
          },
          {
            "cookieAuth": []
          }
        ],
        "tags": [
          "Authentication"
        ],
        "summary": "Retrieve authentication",
        "description": "Retrieve operation in the Authentication module. Organization scope is enforced from the authenticated session."
      }
    },
...
  },
  "info": {
    "title": "ORBIT Fleet API",
    "description": "Complete fleet operations API: authentication, vehicles, TCP-normalized telemetry, drivers, geofences, alerts, maintenance, fuel, tasks, device commands, cameras, configuration, users, audit and reports.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access token returned by POST /v1/auth/login"
      },
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "access_token",
        "description": "HttpOnly browser session cookie"
      }
    },
    "schemas": {
      "LoginDto": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "admin@orbit.sa",
            "description": "User email address"
          },
          "password": {
            "type": "string",
            "format": "password",
            "minLength": 8,
            "example": "Orbit@2026",
            "description": "Account password"
          }
        }
      },
      "RefreshDto": {
        "type": "object",
        "properties": {
          "refreshToken": {
            "type": "string",
            "example": "eyJhbGciOiJIUzI1Ni...",
            "description": "Refresh JWT; optional when the HttpOnly refresh cookie exists"
          }
        }
      },
...
    }
  }
}